+++ include/main.hpp
@@ -5,6 +5,7 @@
 #include <memory_debug_message_info.hpp>
 #include <microros.hpp>
 #include <map>
+#include "MultiDistanceSensor.h"
 
 static UARTSerial microros_serial(RPI_SERIAL_TX, RPI_SERIAL_RX);
