--- src/rosbot_sensors.cpp
@@ -1,8 +1,0 @@
-#include "rosbot_sensors.h"
-
-namespace rosbot_sensors{
-
-
-
-
-}
